Refactor: separate challenge reducer logic as thunk and selector
Modify: get prevContainer from Droggable in addChildTree

Modify: remove duplicate find subChallenge util in selectData
